x/mirror/keeper/grpc_query_erc_20_token_pairs.go
package keeper

import (
	"context"

	sdk "github.com/cosmos/cosmos-sdk/types"
	"github.com/zeta-chain/zetacore/x/mirror/types"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/status"
)

func (k Keeper) ERC20TokenPairs(c context.Context, req *types.QueryGetERC20TokenPairsRequest) (*types.QueryGetERC20TokenPairsResponse, error) {
	if req == nil {
		return nil, status.Error(codes.InvalidArgument, "invalid request")
	}
	ctx := sdk.UnwrapSDKContext(c)

	val, found := k.GetERC20TokenPairs(ctx)
	if !found {
		return nil, status.Error(codes.NotFound, "not found")
	}

	return &types.QueryGetERC20TokenPairsResponse{ERC20TokenPairs: val}, nil
}
